feat: 64-bit-offset strings end-to-end (LargeUtf8) + byte-bounded merge materialization
Wide-TEXT tables overflow 32-bit Utf8 offset arrays (2GiB value cap per
array — 'Offset overflow' at i32::MAX) anywhere a decoded batch or a
whole-merge concat materializes them. Three coordinated changes:

1. Iceberg string now converts to LargeUtf8 (Binary already mapped to
   LargeBinary). The reader widens Utf8/Binary to their 64-bit forms via
   the arrow schema-hint mechanism (same as INT96 coercion), matched by
   field-id so variant sub-fields keep their exact physical types; decode
   materializes 64-bit offsets directly. Parquet physical bytes are
   unchanged (BYTE_ARRAY either way). Constant/literal builders, the
   positional-delete loader, string-leaf extraction, inspect metadata
   tables and the transformer's helpers all gain the wide forms; foreign
   32-bit sources cast at insert/merge coercion (pinned in tests).

2. The MoR merge write path is byte-bounded: consume-loop chunks derive
   rows from a byte target against measured row width (row count is only
   a ceiling); the evaluated-SET store stays CHUNKED for the merge's
   lifetime — (clause, chunk, row) addressing replaces the whole-merge
   per-column concat; the late-materialization fetch sizes decode batches
   from footer row-group byte stats (free — footers already read).
   MorMergeOptions::chunk_target_bytes overrides the 32MiB default.

3. merge_into reports peak_mem_bytes: a PeakTrackingPool wraps the
   session memory pool (bounded or unbounded) and records the pool-wide
   reserved high-water mark; TrackConsumersPool alone drops per-consumer
   peaks on unregister, so an end-of-query read would miss finished
   operators. Env knob MERGE_MOR_CHUNK_TARGET_MB plumbs the byte target.

New tests: large-offset widening visitor (3), byte-bounded chunk sizing
(3), a cross-chunk SET-store + LargeUtf8 end-to-end merge (exact per-row
payload asserts), insert-path cast pinning. Full suites green: 1757
tests across core, datafusion integration, compaction and bindings.

+++ crates/iceberg/src/arrow/reader/projection.rs
@@ -1156,7 +1156,7 @@ message schema {
         assert_eq!(batch.num_columns(), 2);
 
         // Verify position-based mapping: field_id 1 → position 0, field_id 2 → position 1
-        let name_array = batch.column(0).as_string::<i32>();
+        let name_array = batch.column(0).as_string::<i64>();
         assert_eq!(name_array.value(0), "Alice");
         assert_eq!(name_array.value(1), "Bob");
         assert_eq!(name_array.value(2), "Charlie");
@@ -1259,7 +1259,7 @@ message schema {
         assert_eq!(batch.num_columns(), 2);
 
         // field 2 (`name`) must come from physical column 0, not be NULL-filled.
-        let name_array = batch.column(0).as_string::<i32>();
+        let name_array = batch.column(0).as_string::<i64>();
         assert_eq!(
             name_array.null_count(),
             0,
@@ -1270,7 +1270,7 @@ message schema {
         assert_eq!(name_array.value(2), "Charlie");
 
         // field 4 (`subdept`) must come from physical column 1.
-        let subdept_array = batch.column(1).as_string::<i32>();
+        let subdept_array = batch.column(1).as_string::<i64>();
         assert_eq!(subdept_array.null_count(), 0);
         assert_eq!(subdept_array.value(0), "comms");
         assert_eq!(subdept_array.value(1), "tax");
@@ -1369,9 +1369,9 @@ message schema {
         );
 
         let batch = &result[0];
-        let name_array = batch.column(0).as_string::<i32>();
+        let name_array = batch.column(0).as_string::<i64>();
         assert_eq!(name_array.value(0), "Alice");
-        let subdept_array = batch.column(1).as_string::<i32>();
+        let subdept_array = batch.column(1).as_string::<i64>();
         assert_eq!(subdept_array.value(0), "Bob");
     }
 
@@ -1459,11 +1459,11 @@ message schema {
         assert_eq!(batch.num_rows(), 2);
         assert_eq!(batch.num_columns(), 2);
 
-        let col1_array = batch.column(0).as_string::<i32>();
+        let col1_array = batch.column(0).as_string::<i64>();
         assert_eq!(col1_array.value(0), "a");
         assert_eq!(col1_array.value(1), "b");
 
-        let col3_array = batch.column(1).as_string::<i32>();
+        let col3_array = batch.column(1).as_string::<i64>();
         assert_eq!(col3_array.value(0), "c");
         assert_eq!(col3_array.value(1), "d");
     }
@@ -1546,7 +1546,7 @@ message schema {
         assert_eq!(batch.num_rows(), 2);
         assert_eq!(batch.num_columns(), 3);
 
-        let name_array = batch.column(0).as_string::<i32>();
+        let name_array = batch.column(0).as_string::<i64>();
         assert_eq!(name_array.value(0), "Alice");
         assert_eq!(name_array.value(1), "Bob");
 
@@ -1557,7 +1557,7 @@ message schema {
         assert_eq!(age_array.value(1), 25);
 
         // Verify missing column filled with NULLs
-        let city_array = batch.column(2).as_string::<i32>();
+        let city_array = batch.column(2).as_string::<i64>();
         assert_eq!(city_array.null_count(), 2);
         assert!(city_array.is_null(0));
         assert!(city_array.is_null(1));
@@ -1648,7 +1648,7 @@ message schema {
         let mut all_values = Vec::new();
 
         for batch in &result {
-            let name_array = batch.column(0).as_string::<i32>();
+            let name_array = batch.column(0).as_string::<i64>();
             let value_array = batch
                 .column(1)
                 .as_primitive::<arrow_array::types::Int32Type>();
@@ -1786,7 +1786,7 @@ message schema {
         let person_array = batch.column(1).as_struct();
         assert_eq!(person_array.num_columns(), 2);
 
-        let name_array = person_array.column(0).as_string::<i32>();
+        let name_array = person_array.column(0).as_string::<i64>();
         assert_eq!(name_array.value(0), "Alice");
         assert_eq!(name_array.value(1), "Bob");
 
@@ -2139,7 +2139,7 @@ message schema {
         assert_eq!(id_col.value(2), 9);
         assert_eq!(id_col.value(3), 13);
 
-        let name_col = batch.column(1).as_string::<i32>();
+        let name_col = batch.column(1).as_string::<i64>();
         assert_eq!(name_col.value(0), "Alice");
         assert_eq!(name_col.value(1), "Bob");
         assert_eq!(name_col.value(2), "Charlie");
